+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Added: New function to show an object in an interactive UI (Show-Object)
 * Added: System information on new remoting connection
 * Changed: Update and fix the Workspace commands to only support the Project Manager extension (BREAKING)